+++ pkg/test/framework/features/features.yaml
@@ -147,8 +147,7 @@ features:
       cluster_local:
       multimaster:
       remote:
-    external:
-      externalcontrolplane:
+    istiodremote:
     helm:
       default:
         install:
