+++ tests/integration/istiodremote/testdata/gateway.yaml
